macOS ver bumped to 10.12, use clock_gettime, streamline time.c
