+++ README.md
@@ -23,7 +23,7 @@
 | ---------- | -------------------------- |
 | Windows    | 7                          |
 | Android    | API 26 (8.0)               |
-| macOS      | 10.11                      |
+| macOS      | 10.12                      |
 | iOS / tvOS | *Coming soon!*             |
 | Linux      | `*`                        |
 | Web        | `**` Chrome 86, Firefox 79 |
