--- src/unix/apple/sleep.h
@@ -1,1 +1,0 @@
-../linux/sleep.h
